feat(chalice): top-autocomplete for session's attributes using same endpoint for pa-properties feat(chalice): autocomplete for session's attributes using same endpoint for pa-properties
